Seed db and add validations

+++ MoonReads/Controllers/AuthorController.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
 using MoonReads.Dto;
 using MoonReads.Interfaces;
 using MoonReads.Models;
@@ -10,6 +11,8 @@ namespace MoonReads.Controllers
     [ApiController]
     public class AuthorController : Controller
     {
+        private const string DefaultDescription = "Ten autor nie ma jeszcze opisu";
+        private const string DefaultImageUrl = "https://cdn-icons-png.flaticon.com/512/10701/10701484.png";
         private readonly IAuthorRepository _authorRepository;
         private readonly IMapper _mapper;
 
@@ -55,7 +58,7 @@ namespace MoonReads.Controllers
             if (!_authorRepository.AuthorExists(authorId))
                 return NotFound();
 
-            var books = _mapper.Map<List<BookDto>>(_authorRepository.GetBookByAuthor(authorId));
+            var books = _mapper.Map<List<BookDetailDto>>(_authorRepository.GetBookByAuthor(authorId));
 
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
@@ -83,6 +86,16 @@ namespace MoonReads.Controllers
 
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+
+            if (authorCreate.Description.IsNullOrEmpty())
+            {
+                authorCreate.Description = DefaultDescription;
+            }
+
+            if (authorCreate.ImageUrl.IsNullOrEmpty())
+            {
+                authorCreate.ImageUrl = DefaultImageUrl;
+            }
 
             var authorMap = _mapper.Map<Author>(authorCreate);
 
@@ -128,6 +141,7 @@ namespace MoonReads.Controllers
         [ProducesResponseType(400)]
         [ProducesResponseType(204)]
         [ProducesResponseType(404)]
+        [ProducesResponseType(409)]
         public IActionResult DeleteAuthor(int authorId)
         {
             if (!_authorRepository.AuthorExists(authorId))
@@ -137,6 +151,9 @@ namespace MoonReads.Controllers
 
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+
+            if (_authorRepository.HasBooks(author))
+                return Conflict("Author cannot be deleted because have associated books.");
 
             if (!_authorRepository.DeleteAuthor(author))
             {
